+++ R/base.R
@@ -56,7 +56,8 @@ base64enc <- function(x, convert = TRUE) .Call(secretbase_base64enc, x, convert)
 #'     unserialize back to the original object.
 #'
 #' @return A character string, raw vector, or other object depending on the
-#'     value of \sQuote{convert}.
+#'     value of \sQuote{convert}. If conversion to a character string fails,
+#'     a raw vector will be returned instead (along with a warning).
 #'
 #' @details The value of \sQuote{convert} should be set to TRUE, FALSE or NA to
 #'     be the reverse of the 3 encoding operations (for strings, raw vectors and