+++ NEWS.md
@@ -1,5 +1,7 @@
 # secretbase 1.0.2.9000 (development)
 
+* `base64dec()` now emits a suppressable warning when failing to convert back to a character string.
+
 # secretbase 1.0.2
 
 * Improves hash performance in most situations, especially for large files, by optimizing buffer sizes.